// demo-front/typechain-types/RandomAirdrop.ts
export type BigNumberStruct = { val: BytesLike; bitlen: BigNumberish }

export type BigNumberStructOutput = [val: string, bitlen: BigNumber] & {
    val: string
    bitlen: BigNumber
}
export interface SettedUpValues {
    T: string
    n: string
    nl: string
    g: string
    gl: string
    h: string
    hl: string
    commitDuration: string
    commitRevealDuration: string
    setUpTime: string
}

export declare namespace ICommitRevealRecoverRNG {
    export type CommitRevealValueStruct = {
        c: BigNumberStruct
        a: BigNumberStruct
        participantAddress: BytesLike
    }

    export type CommitRevealValueStructOutput = [
        c: BigNumberStructOutput,
        a: BigNumberStructOutput,
        participantAddress: string
    ] & {
        c: BigNumberStructOutput
        a: BigNumberStructOutput
        participantAddress: string
    }

    export type SetUpValueAtRoundStruct = {
        setUpTime: BigNumberish
        commitDuration: BigNumberish
        commitRevealDuration: BigNumberish
        T: BigNumberish
        proofSize: BigNumberish
        n: BigNumberStruct
        g: BigNumberStruct
        h: BigNumberStruct
    }

    export type SetUpValueAtRoundStructOutput = [
        setUpTime: BigNumber,
        commitDuration: BigNumber,
        commitRevealDuration: BigNumber,
        T: BigNumber,
        proofSize: BigNumber,
        n: BigNumberStructOutput,
        g: BigNumberStructOutput,
        h: BigNumberStructOutput
    ] & {
        setUpTime: BigNumber
        commitDuration: BigNumber
        commitRevealDuration: BigNumber
        T: BigNumber
        proofSize: BigNumber
        n: BigNumberStructOutput
        g: BigNumberStructOutput
        h: BigNumberStructOutput
    }

    export type UserAtRoundStruct = {
        index: BigNumberish
        committed: boolean
        revealed: boolean
    }

    export type UserAtRoundStructOutput = [
        index: BigNumber,
        committed: boolean,
        revealed: boolean
    ] & { index: BigNumber; committed: boolean; revealed: boolean }

    export type ValueAtRoundStruct = {
        numOfParticipants: BigNumberish
        count: BigNumberish
        bStar: BytesLike
        commitsString: BytesLike
        omega: BigNumberStruct
        stage: BigNumberish
        isCompleted: boolean
        isAllRevealed: boolean
    }

    export type ValueAtRoundStructOutput = [
        numOfParticipants: BigNumber,
        count: BigNumber,
        bStar: string,
        commitsString: string,
        omega: BigNumberStructOutput,
        stage: BigNumber,
        isCompleted: boolean,
        isAllRevealed: boolean
    ] & {
        numOfParticipants: BigNumber
        count: BigNumber
        bStar: string
        commitsString: string
        omega: BigNumberStructOutput
        stage: BigNumber
        isCompleted: boolean
        isAllRevealed: boolean
    }

    export type VDFClaimStruct = {
        T: BigNumberish
        x: BigNumberStruct
        y: BigNumberStruct
        v: BigNumberStruct
    }

    export type VDFClaimStructOutput = [
        T: BigNumber,
        x: BigNumberStructOutput,
        y: BigNumberStructOutput,
        v: BigNumberStructOutput
    ] & {
        T: BigNumber
        x: BigNumberStructOutput
        y: BigNumberStructOutput
        v: BigNumberStructOutput
    }
}